refactor(deployments): more refactoring to move related pieces together
