+++ src/app/space/create/deployments/appenvs/deployment-card.component.ts
@@ -7,8 +7,8 @@ import {
 
 import { Observable } from 'rxjs';
 
-import { DeploymentsService } from '../../services/deployments.service';
-import { Environment } from '../../models/environment';
+import { DeploymentsService } from '../services/deployments.service';
+import { Environment } from '../models/environment';
 
 // Makes patternfly charts available
 import 'patternfly/dist/js/patternfly-settings.js';
