+++ src/app/space/create/deployments/resource-usage/resource-card.component.ts
