+++ src/app/space/create/deployments/resource-usage/utilization-bar.component.ts
